fix circular depdendency
Moved the notify message function to a better location to avoid
circular dependency
